% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_usk.R
\encoding{UTF-8}
\name{plot_usk}
\alias{plot_usk}
\title{Unbalanced Scott-Knott (USK) on a Ggplot2 Graphic}
\usage{
plot_usk(test)
}
\arguments{
\item{test}{is the \code{\link{plot_usk}} function.}
}
\value{
This function returns a ggplot2 graph with:
\itemize{
\item \code{title}: "Scott-Knott";
\item \code{x}:  treatments;
\item \code{y}:   average value of 'observation' for each treatment;
\item \code{geom_errorbar}: lowest to higher values of 'observation' for each treatment;
\item \code{color}: which group the treatment was classified.
}
}
\description{
\code{\link{plot_usk}} is used to show the \code{\link{usktest}} result,
that do the Scott-Knott cluster analysis (1974) for unbalanced designs proposed at 2017
(CONRADO, Thiago Vincenzi et al), in a ggplot2 graphic.
To learn more, see the \href{http://ref.scielo.org/ws792m}{article}.
}
\examples{

taus=c(4,4,-4,-4,9,-9)
Tratamento<-as.factor(rep(c(paste("trat",seq(1:length(taus)))),3))
erro<-rnorm(3*length(taus),0,1)
y<-2+taus+erro
y[round(runif(1,min=1,max=length(y)),0)]<-NA
dados<-data.frame(y,Tratamento)

test<-usktest(y~Tratamento,dados)
#plotly
plotly_usk(test)
#ggplot2
plot_usk(test)

}
\references{
CONRADO, T. V; FERREIRA, D. F.; SCAPIM, C. A.; MALUF, W. R. (2017) \emph{Adjusting the Scott-Knott cluster analyses for unbalanced designs}.
}
\seealso{
For another graphic: \code{plotly_usk}. See also the packages \href{https://cran.r-project.org/web/packages/ScottKnott/index.html}{ScottKnott}, \href{https://cran.r-project.org/web/packages/ScottKnottESD/index.html}{ScottKnottESD} and \href{https://cran.r-project.org/web/packages/multcomp/index.html}{multcomp}.
}
\author{
Felipe Ferreira \email{felipe179971@hotmail.com}
}
