% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/RUNscott.R
\name{RUNscott}
\alias{RUNscott}
\title{...}
\usage{
RUNscott(dataset, var1, var2, alpha, v, MSE)
}
\arguments{
\item{dataset}{...}

\item{var1}{...}

\item{var2}{...}

\item{alpha}{...}

\item{v}{...}

\item{MSE}{...}
}
\description{
...
}
\author{
Felipe Ferreira
}
