% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usktest.R
\name{usktest}
\alias{usktest}
\title{...}
\usage{
usktest(dataset, var1, var2, alpha, graphic, ANOVA)
}
\arguments{
\item{dataset}{...}

\item{var1}{...}

\item{var2}{...}

\item{alpha}{...}

\item{graphic}{...}

\item{ANOVA}{...}
}
\description{
...
}
