% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_usk.R
\encoding{UTF-8}
\name{plot_usk}
\alias{plot_usk}
\title{U}
\usage{
plot_usk(test)
}
\arguments{
\item{test}{...}
}
\value{
This function returns the ANOVA table, a graph and a \code{data.frame} including columns:
\itemize{
\item '\code{treatments}': treatments.
\item \code{group}: which group the treatment was classified.
\item \code{mean}: average value of 'observation' for each treatment.
\item \code{min}: lowest value of 'observation' for each treatment.
\item \code{max}: higher value of 'observation' for each treatment.
}
}
\description{
\code{\link{usktest}} is a function used to do the Scott-Knott
cluster analyses (1974) for unbalanced designs proposed at 2017 (CONRADO, Thiago
Vincenzi et al). To learn more, see the \href{http://ref.scielo.org/ws792m}{article}.
}
\author{
Felipe Ferreira \email{felipe179971@hotmail.com}
}
