% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usktest.R
\encoding{UTF-8}
\name{usktest}
\alias{usktest}
\title{Unbalanced Scott-Knott (USK)}
\usage{
usktest(formula, dataset, alpha = 0.05, graphic = TRUE, ANOVA = TRUE)
}
\arguments{
\item{formula}{at the moment, this package only does the Unbalanced Scott-Knott for single factor analysis of variance, so your 'formula' must be \code{observation ~ treatment}}

\item{dataset}{an indication to the database being used.}

\item{alpha}{type I error the researcher can accept (the default is 0.05).}

\item{graphic}{if TRUE (the default), print a chart ggplotly.}

\item{ANOVA}{if TRUE (the default), print an ANOVA table.}
}
\value{
This function returns the ANOVA table, a graph and a \code{data.frame} including columns:
\itemize{
\item '\code{treatments}': treatments.
\item \code{group}: which group the treatment was classified.
\item \code{mean}: average value of 'observation' for each treatment.
\item \code{min}: lowest value of 'observation' for each treatment.
\item \code{max}: higher value of 'observation' for each treatment.
}
}
\description{
\code{\link{usktest}} is a function used to do the Scott-Knott
cluster analyses (1974) for unbalanced designs proposed at 2017 (CONRADO, Thiago
Vincenzi et al). To learn more, see the \href{http://ref.scielo.org/ws792m}{article}.
}
\examples{

taus=c(4,4,-4,-4,9,-9)
Tratamento<-as.factor(rep(c(paste("trat",seq(1:length(taus)))),3))
erro<-rnorm(3*length(taus),0,1)
y<-2+taus+erro
y[round(runif(1,min=1,max=length(y)),0)]<-NA
dados<-data.frame(y,Tratamento)

usktest(y~Tratamento,dados)

}
\author{
Felipe Ferreira \email{felipe179971@hotmail.com}
}
