% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usktest.R
\encoding{UTF-8}
\name{usktest}
\alias{usktest}
\title{Unbalanced Scott-Knott (USK)}
\usage{
usktest(formula, dataset, alpha = 0.05, ANOVA = TRUE)
}
\arguments{
\item{formula}{at the moment, this package only does the Unbalanced Scott-Knott for single factor analysis of variance, so your 'formula' must be \code{observation ~ treatment}}

\item{dataset}{an indication to the database being used.}

\item{alpha}{type I error the researcher can accept (the default is 0.05).}

\item{ANOVA}{if TRUE (the default), print an ANOVA table.}
}
\value{
If \code{ANOVA=TRUE}, this function returns the ANOVA table and the Scott-Knott table that are a \code{data.frame} including 5 columns:
\itemize{
\item '\code{treatments}': the name of the 'treatments' parameter.
\item \code{group}: which group the treatment was classified.
\item \code{mean}: average value of 'observation' for each treatment.
\item \code{min}: lowest value of 'observation' for each treatment.
\item \code{max}: higher value of 'observation' for each treatment.
}
}
\description{
\code{\link{usktest}} is a function used to do the Scott-Knott
cluster analysis (1974) for unbalanced designs proposed at 2017 (CONRADO, T. V; FERREIRA, D. F.; SCAPIM, C. A.; MALUF, W. R.).
To learn more, see the \href{http://ref.scielo.org/ws792m}{article}.
}
\examples{

set.seed(3)
taus=c(4,4,-4,-4,9,-9)
Tratamento<-as.factor(rep(c(paste("trat",seq(1:length(taus)))),3))
erro<-rnorm(3*length(taus),0,1)
y<-2+taus+erro
y[round(runif(1,min=1,max=length(y)),0)]<-NA
dados<-data.frame(y,Tratamento)

test<-usktest(y~Tratamento,dados)

#ggplot2
plot_usk(test)
#plotly
#plotly_usk(test)


}
\references{
CONRADO, T. V; FERREIRA, D. F.; SCAPIM, C. A.; MALUF, W. R. (2017) \emph{Adjusting the Scott-Knott cluster analyses for unbalanced designs}.
}
\seealso{
For graphics: \code{plot_usk} or \code{plotly_usk}. See also the packages \href{https://cran.r-project.org/web/packages/ScottKnott/index.html}{ScottKnott}, \href{https://cran.r-project.org/web/packages/ScottKnottESD/index.html}{ScottKnottESD} and \href{https://cran.r-project.org/web/packages/multcomp/index.html}{multcomp}.
}
\author{
Felipe Ferreira \email{felipe179971@hotmail.com}
}
