% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usktest.R
\encoding{UTF-8}
\name{usktest}
\alias{usktest}
\title{Unbalanced Scott-Knott (USK)}
\usage{
usktest(dataset, var1, var2, alpha, graphic, ANOVA)
}
\arguments{
\item{dataset}{An indication to the database being used.}

\item{var1}{A quantitative variable containing the metric results of the experiment.}

\item{var2}{A factor variable indicating the treatments.}

\item{alpha}{Type I error the researcher can accept (the default is 0.05).}

\item{graphic}{If TRUE (the default), export a chart ggplotly.}

\item{ANOVA}{If TRUE (the default), export an ANOVA table.}
}
\value{
This function returns the ANOVA table, a graph and a \code{data.frame} including columns:
\itemize{
\item "var2": Treatments.
\item group: Which group the treatment was classified.
\item mean: Average value of "var1" for each treatment.
\item min: Lowest value of "var1" for each treatment.
\item max Higher value of "var1" for each treatment.
}
}
\description{
\code{\link{usktest}} is a function used to do the Scott-Knott
cluster analyses (1974) for unbalanced designs proposed at 2017 (CONRADO, Thiago
Vincenzi et al). To learn more, see the \href{http://ref.scielo.org/ws792m}{article}
}
\examples{

taus=c(4,4,-4,-4,9,-9)
Tratamento<-as.factor(rep(c(paste("trat",seq(1:length(taus)))),3))
erro<-rnorm(3*length(taus),0,1)
y<-2+taus+erro
y[round(runif(1,min=1,max=length(y)),0)]<-NA
dados<-data.frame(y,Tratamento)

usktest(dados,"y","Tratamento",0.05,graphic=TRUE,ANOVA=TRUE)

}
\author{
Felipe Ferreira \email{felipe179971@hotmail.com}
}
