% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/usktest.R
\encoding{UTF-8}
\name{usktest}
\alias{usktest}
\title{Unbalanced Scott-Knott (USK)}
\usage{
usktest(formula, dataset, alpha = 0.05, graphic = TRUE, ANOVA = TRUE)
}
\arguments{
\item{formula}{a quantitative variable containing the metric results of the experiment.}

\item{dataset}{an indication to the database being used.}

\item{alpha}{type I error the researcher can accept (the default is 0.05).}

\item{graphic}{if TRUE (the default), export a chart ggplotly.}

\item{ANOVA}{if TRUE (the default), export an ANOVA table.}
}
\value{
This function returns the ANOVA table, a graph and a \code{data.frame} including columns:
\itemize{
\item "var2": Treatments.
\item group: Which group the treatment was classified.
\item mean: Average value of "var1" for each treatment.
\item min: Lowest value of "var1" for each treatment.
\item max Higher value of "var1" for each treatment.
}
}
\description{
\code{\link{usktest}} is a function used to do the Scott-Knott
cluster analyses (1974) for unbalanced designs proposed at 2017 (CONRADO, Thiago
Vincenzi et al). To learn more, see the \href{http://ref.scielo.org/ws792m}{article}.
}
\examples{

taus=c(4,4,-4,-4,9,-9)
Tratamentoo<-as.factor(rep(c(paste("trat",seq(1:length(taus)))),3))
erro<-rnorm(3*length(taus),0,1)
yy<-2+taus+erro
yy[round(runif(1,min=1,max=length(yy)),0)]<-NA
dados<-data.frame(yy,Tratamentoo)

usktest(yy~Tratamentoo,dados)

}
\author{
Felipe Ferreira \email{felipe179971@hotmail.com}
}
